hapus selamanya dan aktifkan kembali

+++ fungsi/aksi.php
@@ -389,9 +389,71 @@ if ($lihat == 'data_peminjaman_kunci' && $metode == 'hapus') {
     }
 
     // delete
+    $sql = "UPDATE pinjam_kunci SET terhapus='1' WHERE id='$id'";
+    if ($connectdb->query($sql) === TRUE) {
+        $_SESSION['success_text'] = array('Data peminjaman kunci berhasil dihapus. Data masih tersimpan dengan status dihapus.');
+    } else {
+        error_log('Error hapus data peminjaman kunci. ' . $connectdb->error);
+        $_SESSION['error_text'] = array('Data tidak bisa dihapus. Kegagalan sistem. Silahkan Coba lagi!');
+    }
+
+    header($location_header);
+    die();
+}
+
+/**
+ * Hapus Selamanya Data Peminjaman Kunci
+ * ============================================================================
+ */
+if ($lihat == 'data_peminjaman_kunci' && $metode == 'hapus_selamanya') {
+    global $id;
+
+    $location_header = 'Location: ' . $config['base_url'] . '/admin?lihat=data_peminjaman_kunci';
+
+    // check exists
+    $sql = "SELECT * FROM pinjam_kunci WHERE id='$id'";
+    $hasil = $connectdb->query($sql);
+    if ($hasil->num_rows < 1) {
+        $_SESSION['error_text'] = array('Data peminjaman kunci tidak ditemukan.');
+        header($location_header);
+        die();
+    }
+
+    // delete
     $sql = "DELETE FROM pinjam_kunci WHERE id='$id'";
     if ($connectdb->query($sql) === TRUE) {
-        $_SESSION['success_text'] = array('Data peminjaman kunci berhasil dihapus.');
+        $_SESSION['success_text'] = array('Data peminjaman kunci berhasil dihapus selamanya.');
+    } else {
+        error_log('Error hapus data peminjaman kunci. ' . $connectdb->error);
+        $_SESSION['error_text'] = array('Data tidak bisa dihapus. Kegagalan sistem. Silahkan Coba lagi!');
+    }
+
+    header($location_header);
+    die();
+}
+
+/**
+ * Hapus Selamanya Data Peminjaman Kunci
+ * ============================================================================
+ */
+if ($lihat == 'data_peminjaman_kunci' && $metode == 'aktifkan') {
+    global $id;
+
+    $location_header = 'Location: ' . $config['base_url'] . '/admin?lihat=data_peminjaman_kunci';
+
+    // check exists
+    $sql = "SELECT * FROM pinjam_kunci WHERE id='$id'";
+    $hasil = $connectdb->query($sql);
+    if ($hasil->num_rows < 1) {
+        $_SESSION['error_text'] = array('Data peminjaman kunci tidak ditemukan.');
+        header($location_header);
+        die();
+    }
+
+    // update
+    $sql = "UPDATE pinjam_kunci SET terhapus='0' WHERE id='$id'";
+    if ($connectdb->query($sql) === TRUE) {
+        $_SESSION['success_text'] = array('Data peminjaman kunci berhasil diaktifkan kembali.');
     } else {
         error_log('Error hapus data peminjaman kunci. ' . $connectdb->error);
         $_SESSION['error_text'] = array('Data tidak bisa dihapus. Kegagalan sistem. Silahkan Coba lagi!');
